separate logic and ui
