+++ components/register/RegisterContainer.js
@@ -1,5 +1,5 @@
-import React, {useState} from 'react'
-import {View, Text, TextInput, TouchableOpacity} from 'react-native'
+import React, {useState} from 'react';
+import RegisterComponent from './RegisterComponent';
 
 const Register = ({register, serverMsg, isLoggedIn, isRegisterVisible, showComponent}) => {
     const [email, setEmail] = useState('');
@@ -13,19 +13,11 @@ const Register = ({register, serverMsg, isLoggedIn, isRegisterVisible, showCompo
         else if(changed.password1)
             setPassword1(changed.password1);
         else if(changed.password2)
-            setPassword2(changed.password2)
+            setPassword2(changed.password2);
         else if(changed.age)
-            setAge(changed.age)
+            setAge(changed.age);
         else
             console.log("something's wrong in ./components/register/registerContainer, handleChange method.");
-
-        // changed.email ? 
-        //     setEmail(changed.email) : 
-        //     changed.password1 ? 
-        //         setPassword1(changed.password1) : 
-        //         changed.password2 ?
-        //             setPassword2(changed.password2) :
-        //             setAge(changed.age);
     }
 
     function handlePress(btnText){
@@ -36,68 +28,14 @@ const Register = ({register, serverMsg, isLoggedIn, isRegisterVisible, showCompo
         return null;
 
     return(
-        <View>
-            <Text>{serverMsg}</Text>
-            <TextInput 
-                name='email' 
-                placeholder={'Enter email'}
-                style={styles.emailInput}
-                onChangeText={(value) => handleChange({email: value})}
-                />
-            <TextInput 
-                name='password1' 
-                placeholder={'Enter password'}
-                style={styles.passwordInput}
-                onChangeText={(value) => handleChange({password1: value})}
-            />
-            <TextInput 
-                name='password2' 
-                placeholder={'Enter password again'}
-                style={styles.passwordInput}
-                onChangeText={(value) => handleChange({password2: value})}
-            />
-            <TextInput 
-                name='age' 
-                placeholder={'Age'}
-                style={styles.ageInput}
-                onChangeText={(value) => handleChange({age: parseInt(value)})}
-            />
-            <TouchableOpacity
-                onPress={() => handlePress('Register')}
-                style={styles.button}
-            >
-                <Text style={styles.buttonText}>Register</Text>
-            </TouchableOpacity>
-            <TouchableOpacity
-                onPress={() => handlePress('Login')}
-                style={styles.button}
-            >
-                <Text style={styles.buttonText}>Have an account?</Text>
-            </TouchableOpacity>
-        </View>
+        <RegisterComponent
+            handleChange={handleChange}
+            handlePress={handlePress}
+            isLoggedIn={isLoggedIn}
+            serverMsg={serverMsg}
+            isRegisterVisible={isRegisterVisible}
+        />
     )
 }
 
-const styles = {
-    emailInput: {
-
-    },
-
-    passwordInput: {
-
-    },
-    ageInput: {
-
-    },
-    button: {
-        margin: 2,
-        padding: 5,
-        alignItems: 'center',
-        backgroundColor: 'darkslateblue'
-    },
-    buttonText: {
-        color: 'white'
-    }
-}
-
-export default Register
+export default Register;
